[#34653,#34655] pass more tests on windows, isolate more subroutine calling external libs

See merge request codeaster/src!668

+++ bibfor/include/asterc/gtopti.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@
 interface
     subroutine gtopti(opt, vali, iret)
         character(len=*) :: opt
-        integer :: vali
-        integer :: iret
+        integer(kind=8) :: vali
+        integer(kind=8) :: iret
     end subroutine gtopti
 end interface
