+++ bibfor/include/asterfort/asmpi_comm_point.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2020 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -23,13 +23,13 @@ interface
                                 scr)
         character(len=*), intent(in) :: optmpi
         character(len=*), intent(in) :: typsca
-        integer, intent(in) :: nudest
-        integer, intent(in) :: numess
-        integer, intent(in), optional :: nbval
-        integer, intent(inout), optional :: vi(*)
+        integer(kind=8), intent(in) :: nudest
+        integer(kind=8), intent(in) :: numess
+        integer(kind=8), intent(in), optional :: nbval
+        integer(kind=8), intent(inout), optional :: vi(*)
         integer(kind=4), intent(inout), optional :: vi4(*)
         real(kind=8), intent(inout), optional :: vr(*)
-        integer, intent(inout), optional :: sci
+        integer(kind=8), intent(inout), optional :: sci
         integer(kind=4), intent(inout), optional :: sci4
         real(kind=8), intent(inout), optional :: scr
     end subroutine asmpi_comm_point
