+++ bibfor/include/asterfort/copmod.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -26,7 +26,7 @@ interface
         complex(kind=8), optional, intent(out) :: bmodz(*)
         character(len=*), optional, intent(in) :: champ
         character(len=*), optional, intent(in) :: numer
-        integer, optional, intent(in) :: nbmodes
-        integer, optional, intent(in) :: nequa
+        integer(kind=8), optional, intent(in) :: nbmodes
+        integer(kind=8), optional, intent(in) :: nequa
     end subroutine copmod
 end interface
