+++ bibfor/include/asterfort/dcapno.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -22,7 +22,7 @@ interface
     subroutine dcapno(resuz, typchz, iord, chavaz)
         character(len=*) :: resuz
         character(len=*) :: typchz
-        integer :: iord
+        integer(kind=8) :: iord
         character(len=*) :: chavaz
     end subroutine dcapno
 end interface
