+++ bibfor/include/asterfort/elg_calc_solu.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -20,14 +20,14 @@
 interface 
     subroutine elg_calc_solu(matas1, nsecm, rsolu2, rsolu1, omega2, ke_mass)
         character(len=19) :: matas1
-        integer :: nsecm
+        integer(kind=8) :: nsecm
         real(kind=8) :: rsolu2(*)
         real(kind=8) :: rsolu1(*)
         real(kind=8), optional, intent(in)::omega2
 #ifdef ASTER_HAVE_PETSC
         PetscInt, optional, intent(in) :: ke_mass
 #else
-        integer, optional, intent(in) :: ke_mass
+        integer(kind=8), optional, intent(in) :: ke_mass
 #endif
     end subroutine elg_calc_solu
 end interface
