+++ bibfor/include/asterfort/getvid.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2024 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -23,10 +23,10 @@ interface
                       scal, nbret)
         character(len=*), intent(in) :: motfac
         character(len=*), intent(in) :: motcle
-        integer, intent(in), optional :: iocc
-        integer, intent(in), optional :: nbval
+        integer(kind=8), intent(in), optional :: iocc
+        integer(kind=8), intent(in), optional :: nbval
         character(len=*), intent(inout), optional :: vect(*)
         character(len=*), intent(inout), optional :: scal
-        integer, intent(out), optional :: nbret
+        integer(kind=8), intent(out), optional :: nbret
     end subroutine getvid
 end interface
