+++ bibfor/include/asterfort/jecrec.h
@@ -1,5 +1,5 @@
 ! --------------------------------------------------------------------
-! Copyright (C) 1991 - 2017 - EDF R&D - www.code-aster.org
+! Copyright (C) 1991 - 2025 - EDF R&D - www.code-aster.org
 ! This file is part of code_aster.
 !
 ! code_aster is free software: you can redistribute it and/or modify
@@ -25,6 +25,6 @@
               character(len=*), intent(in) :: accelu
               character(len=*), intent(in) :: stoclu
               character(len=*), intent(in) :: longlu
-              integer, intent(in) :: nmax
+              integer(kind=8), intent(in) :: nmax
             end subroutine jecrec
           end interface
